File input is working. Closed Chrome browser fully and tried again and it was working.
