+++ src/components/band/BandList.tsx
@@ -6,6 +6,7 @@ import { useEffect, useState } from "react";
 import BandDelete from "./DeleteBand";
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
+import Image from "next/image";
 
 export default function BandList() {
   const [bands, setBands] = useState<BandData[]>([]);
@@ -30,15 +31,24 @@ export default function BandList() {
 
   return (
     <>
-      <div className="max-h-full min-h-full">
+      <div className="my-8 flex max-h-full min-h-full items-start justify-between gap-4">
         {filteredBands.map((band) => (
-          <li key={band.id}>
-            <Link href={`/bands/${band.id}`}>
-            <h1>{band.bandName}</h1>
-            <div><BandDelete bandId={band.id}/></div>
+          <div className="min-h-24 min-w-56" key={band.id}>
+            <Link className="" href={`/bands/${band.id}`}>
+              <div className="items-between flex-col items-baseline">
+                <h1 className="text-left text-2xl">{band.bandName}</h1>
+                <Image
+                  src={band.bandPic || "/default-image.png"}
+                  alt={band.bandName}
+                  width={50}
+                  height={50}
+                />
+              </div>
+              <div className="flex justify-end align-baseline">
+                <BandDelete bandId={band.id} />
+              </div>
             </Link>
-           
-          </li>
+          </div>
         ))}
       </div>
     </>
